1-D NFFT

1-D Non-Uniform Direct Fourier Tranform

ndft_1d(x, f_hat)

nfft_1d(x, f_hat)

nfft_adjoint_1d(x, f, n)

ndft_adjoint_1d(x, f, n)

Arguments

x
(real) vector of nodes of length m, must be in [-0.5,0.5).

f_hat
(complex) vector of \(\hat{f}\) entries - (of length n) and length must be even.

f
frequencies for adjoint, same length as x

n
number of frequencies for transform, specified for adjoint.

Value

vector of \(f\), the results of the transform (of length m).

Details

The non-uniform Fourier transform takes non-uniform samples \(x\) from the $d$-dimensional torus \([0.5,0.5)^d\).

The NDFT functions compute the Fourier transform directly. This is slow. The NFFT functions use the FFT to compute this, which should be faster. The adjoint, in this case, is not the same as the inverse. Solving the inverse problem requires approximations. Here we present the 1D NDFT, NFFT, and their adjoints. You most likely want to use the nfft_1d and nfft_adjoint_1d functions rather than the dft functions.

Functions

nfft_1d:

nfft_adjoint_1d:

ndft_adjoint_1d:

References

Keiner, J., Kunis, S., and Potts, D. ''Using NFFT 3 - a software library for various nonequispaced fast Fourier transforms'' ACM Trans. Math. Software,36, Article 19, 1-30, 2009.

Examples

set.seed(20190722)
x <- runif(19) - 0.5
#f_hat <- runif(14) + 1i*runif(14)
f_hat = 1:14
for(i in 1:14) f_hat[i] = runif(1)*1i + runif(1)

ndft_1d(x, f_hat)

#>  [1] -0.0456344+0.6552138i  0.7609645-0.2068335i  0.9669170-0.2650461i
#>  [4]  0.3364985+0.8481410i  0.8616118+0.6593409i  0.4079021+0.8492695i
#>  [7]  0.5037352+0.7777328i  4.3022482+2.2288860i -1.0830297+0.6838535i
#> [10]  1.4042922-1.1465600i -0.7130283+0.7064547i -0.1974753+1.5038762i
#> [13]  2.5030605+3.3386115i  1.3162611+1.8288851i -2.7053836+1.5109461i
#> [16]  1.1135791-0.2623614i -1.4009907+1.7914146i  1.4202173+2.4114651i
#> [19] -1.5733905+2.0034500i

nfft_1d(x, f_hat)

#>  [1] -0.0456344+0.6552138i  0.7609645-0.2068335i  0.9669170-0.2650461i
#>  [4]  0.3364985+0.8481410i  0.8616118+0.6593409i  0.4079021+0.8492695i
#>  [7]  0.5037352+0.7777328i  4.3022482+2.2288860i -1.0830297+0.6838535i
#> [10]  1.4042922-1.1465600i -0.7130283+0.7064547i -0.1974753+1.5038762i
#> [13]  2.5030605+3.3386115i  1.3162611+1.8288851i -2.7053836+1.5109461i
#> [16]  1.1135791-0.2623614i -1.4009907+1.7914146i  1.4202173+2.4114651i
#> [19] -1.5733905+2.0034500i

f <- nfft_1d(x, f_hat)
nfft_adjoint_1d(x,f,14)

#>  [1]  7.711190+ 3.593705i  3.236544- 6.312228i  4.254726+ 7.026398i
#>  [4]  5.806141- 2.212952i  9.680618- 0.870162i  5.439639+10.817064i
#>  [7]  1.753873+ 2.593546i  8.178355+19.916740i 10.600357+11.702932i
#> [10] -3.365497+10.304345i 12.773905+ 1.792603i  6.043838+ 6.657434i
#> [13]  3.099985- 4.750190i -5.230434+ 3.819773i

ndft_adjoint_1d(x,f,14)

#>  [1]  7.711190+ 3.593705i  3.236544- 6.312228i  4.254726+ 7.026398i
#>  [4]  5.806141- 2.212952i  9.680618- 0.870162i  5.439639+10.817064i
#>  [7]  1.753873+ 2.593546i  8.178355+19.916740i 10.600357+11.702932i
#> [10] -3.365497+10.304345i 12.773905+ 1.792603i  6.043838+ 6.657434i
#> [13]  3.099985- 4.750190i -5.230434+ 3.819773i