1-D Non-Uniform Fast Cosine/Sine Tranform

Similar to the NFFT, except the cosine/sine transform instead of the FFT. Note that all values of x must be in \([0,1/2)\).

nfct_1d(x, f_hat)

nfct_adjoint_1d(x, f, n)

nfst_1d(x, f_hat)

nfst_adjoint_1d(x, f, n)

Arguments

x
(real) vector of nodes of length m, must be in [-0.5,0.5).

f_hat
(complex) vector of \(\hat{f}\) entries - (of length n) and length must be even.

f
frequencies for adjoint, same length as x

n
number of frequencies for transform, specified for adjoint.

Functions

nfct_adjoint_1d: The adjoint transform

nfst_1d: The sine transform

nfst_adjoint_1d: The adjoint sine transform

Examples

set.seed(20190722)
x <- runif(19) * 0.5
f_hat = runif(14)

nfct_1d(x, f_hat)

#>  [1]  5.4186730  0.6332174  0.3175322  1.8015275  0.9287882  0.1476851
#>  [7] -0.9246638  1.5552194  1.8183815  0.4206213  1.8188406  0.2603326
#> [13]  0.6326914  3.3465612 -0.5457435  0.2805343  0.9557225  0.6917826
#> [19]  0.3048201

nfst_1d(x, f_hat)

#>  [1]  2.4633346  0.4176734 -0.4603488  1.8113051  1.6707539  1.3080388
#>  [7] -0.9226009  1.4518733  1.8787348  0.4107718  1.8660739 -0.5701987
#> [13]  0.7705659  3.9714202 -0.1757409 -0.5442597  0.7553293  0.8006286
#> [19]  2.0681834

f <- nfct_1d(x, f_hat)
g <- nfst_1d(x, f_hat)
nfct_adjoint_1d(x,f,14)

#>  [1] 19.8625235 11.0216384  0.8163386  3.6715299 12.0311126 13.2365001
#>  [7]  0.9603967  2.0420906  3.4180189  9.6718393  7.8085344 -1.6810564
#> [13] -2.5598104  1.9788051

nfst_adjoint_1d(x,g,14) # note the length!

#>  [1] 11.5288234 10.2427723 -1.1227294  0.9195522  9.6304384  9.9918929
#>  [7]  0.5509178  0.7719128  3.5143067  9.2697782 12.0379537  1.2912612
#> [13] -0.1273588